requirements.txt
pvrecorder
numpy
openai
google-auth-oauthlib
aiohttp
PyYAML
langchain
PyDrive
google-api-python-client
chromadb
tiktoken
pandas